Extract shared plugin code
