+++ site/docs/.vuepress/plugins/shared.ts
@@ -1,0 +1,17 @@
+// Adapted from original `code_inline` implementation of markdown-it.
+const HTML_ESCAPE_TEST_RE = /&|<(?!wbr>)|(?<!<wbr)>/;
+const HTML_ESCAPE_REPLACE_RE = /&|<(?!wbr>)|(?<!<wbr)>/g;
+const HTML_REPLACEMENTS: Record<string, string> = {
+  "&": "&amp;",
+  "<": "&lt;",
+  ">": "&gt;",
+  '"': "&quot;",
+};
+function replaceUnsafeChar(ch: string) {
+  return HTML_REPLACEMENTS[ch];
+}
+export function escapeHtml(str: string) {
+  return HTML_ESCAPE_TEST_RE.test(str)
+    ? str.replace(HTML_ESCAPE_REPLACE_RE, replaceUnsafeChar)
+    : str;
+}
